Disable Bootstrap animation and add own open animation
modal(“open”) and modal(“close”) are ignored while a modal is in a
transition state. Sometimes we call open and close while they the
animation is still running. The modal stays open or closed and does not
reflect the state of the app correctly! In the long run we should not
use a modal to display the current state of the app.